backend/requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.23.2
aiohttp>=3.8.5

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.6

# Data Validation
pydantic>=2.4.2
pydantic-settings>=2.0.0
email-validator>=2.0.0

# RAG Components
faiss-cpu>=1.7.4
rank-bm25>=0.2.2
networkx>=3.1
sentence-transformers>=2.2.2
scikit-learn>=1.3.0

# Document Processing
PyPDF2>=3.0.0
python-docx>=1.0.0
markdown>=3.5
python-frontmatter>=1.0.0
PyYAML>=6.0.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.25.0
tenacity>=8.2.3
loguru>=0.7.0
requests>=2.31.0
GitPython>=3.1.40